# swagger.yaml
openapi: 3.0.3
info:
  title: Go Clean Architecture API
  description: API documentation for Auth and Todo features
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                name:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      name:
                        type: string
        '400':
          description: Bad request (invalid format or length)
        '409':
          description: Conflict (email already exists)
  
  /api/v1/auth/login:
    post:
      summary: Login and get JWT token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
        '400':
          description: Bad request
        '401':
          description: Unauthorized (invalid credentials)

  /api/v1/todos:
    post:
      summary: Create a new todo
      tags:
        - Todo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseTodo'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
    
    get:
      summary: List all todos for the authenticated user
      tags:
        - Todo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Todo'
        '401':
          description: Unauthorized

  /api/v1/todos/{id}:
    get:
      summary: Get a specific todo
      tags:
        - Todo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseTodo'
        '401':
          description: Unauthorized
        '404':
          description: Not found (or doesn't belong to user)
    
    put:
      summary: Update a specific todo
      tags:
        - Todo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
                done:
                  type: boolean
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseTodo'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Not found

    delete:
      summary: Delete a specific todo
      tags:
        - Todo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
        '401':
          description: Unauthorized
        '404':
          description: Not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    Todo:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        description:
          type: string
        done:
          type: boolean
    
    SuccessResponseTodo:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Todo'